\name{g.fit}
\alias{g.fit}
\docType{data}
\title{
Mountain Goat Sightability Model Information
}
\description{
Model averaged regression parameters and unconditional variance-covariance matrix for mountain goat sightability model (Rice et al. 2009)
}
\usage{data(g.fit)}
\format{
  The format is:
  beta.g  = list of regression parameters (intercept and parameters associated with GroupSize, Terrain, and X.VegCover) 
  varbeta.g = variance-covariance matrix (associated with beta.g) 
}
\references{
Rice C.G., Jenkins K.J., Chang W.Y. (2009). A Sightability Model for Mountain
Goats. The Journal of Wildlife Management, 73(3), 468-478. 
}
\examples{
data(g.fit)
}
\keyword{datasets}
