\name{print.sightest}
\alias{print.sightest}
\title{Print method for sightability estimator
}
\description{
Prints fitted sightability model, sampling information, and sightability estimate (with confidence interval)
}
\usage{
  \method{print}{sightest}(x,...)
}
\arguments{
  \item{x}{
   Sightability object, output from call to Sight.Est function. 
}
  \item{...}{arguments to be passed to or from other methods}
}
\author{
John Fieberg
}

\seealso{
 \code{\link{Sight.Est}}, \code{\link{summary.sightest}} 
}
\keyword{print}
 
