\name{sampinfo.m}
\alias{sampinfo.m}
\docType{data}
\title{
Data set containing sampling information for observation survey of moose in MN
}
\description{
Data set containing sampling information from a survey of moose in MN (during years 2004-2007)
}
\usage{data(sampinfo.m)}
\format{
  A data frame with 12 observations on the following 5 variables.
  \describe{
    \item{\code{year}}{year of survey}
    \item{\code{stratum}}{stratum identifier}
    \item{\code{Nh}}{number of population units in stratum h}
    \item{\code{nh}}{number of sample units in stratum h}
  }
}
\references{
   Giudice, J H. and Fieberg, J. and Lenarz, M. S.  2012. Spending Degrees of Freedom in a Poor Economy: A Case Study of Building a Sightability Model for Moose in Northeastern Minnesota.
   Journal of Wildlife Management 76(1):75-87.

}
\examples{
data(sampinfo.m)
sampinfo.m
}
\keyword{datasets}
