% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{SightabilityModel-package}
\alias{SightabilityModel-package}
\alias{SightabilityModel}
\title{Wildlife Sightability Modeling}
\description{
Uses logistic regression to model the probability of detection as a function
of covariates.  This model is then used with observational survey data to
estimate population size, while accounting for uncertain detection.  See
Steinhorst and Samuel (1989).
}
\details{
\tabular{ll}{ Package: \tab SightabilityModel\cr Type: \tab Package\cr
Version: \tab 1.3\cr Date: \tab 2014-10-03\cr License: \tab GPL-2\cr
LazyLoad: \tab yes\cr }
}
\references{
Fieberg, J.  2012.  Estimating Population Abundance Using
Sightability Models: R SightabilityModel Package. Journal of Statistical
Software, 51(9), 1-20.  URL http://www.jstatsoft.org/v51/i09/.

Steinhorst, Kirk R. and Samuel, Michael D. 1989. Sightability Adjustment
Methods for Aerial Surveys of Wildlife Populations. Biometrics 45:415--425.
}
\author{
John Fieberg

Maintainer: John Fieberg <jfieberg@umn.edu>
}
\keyword{package}
