% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g.fit}
\alias{g.fit}
\title{Mountain Goat Sightability Model Information}
\format{
The format is: beta.g = list of regression parameters (intercept and
parameters associated with GroupSize, Terrain, and X.VegCover) varbeta.g =
variance-covariance matrix (associated with beta.g)
}
\description{
Model averaged regression parameters and unconditional variance-covariance
matrix for mountain goat sightability model (Rice et al. 2009)
}
\examples{

data(g.fit)

}
\references{
Rice C.G., Jenkins K.J., Chang W.Y. (2009). A Sightability Model
for Mountain Goats. The Journal of Wildlife Management, 73(3), 468-478.
}
\keyword{datasets}
