\name{gdat}
\alias{gdat}
\docType{data}
\title{
Mountain Goat Survey Data from Olympic National park
}
\description{
Mountain Goat Survey Data from Olympic National park collected in 2004
}
\usage{data(gdat)}
\format{
  A data frame with 113 observations on the following 9 variables.
  \describe{
    \item{\code{GroupSize}}{number of animals observed in each independently sighted group [cluster size]}
    \item{\code{Terrain}}{measure of terrain obstruction}
    \item{\code{X.VegCover}}{measure of vegetative obstruction}
    \item{\code{stratum}}{stratum identifier}
    \item{\code{total}}{number of animals observed in each independently sighted group [same as GroupSize]}
    \item{\code{subunit}}{a numeric vector, Plot ID}
  }
}
\source{
Patti Happe (Patti_Happe@nps.gov)
}
\examples{
data(gdat)
}
\keyword{datasets}
