% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdat}
\alias{gdat}
\title{Mountain Goat Survey Data from Olympic National park}
\format{
A data frame with 113 observations on the following 9 variables.
\describe{ 
  \item{GroupSize}{number of animals observed in each
        independently sighted group [cluster size]} 
  \item{Terrain}{measure
        of terrain obstruction} \item{pct.VegCover}{measure of vegetative
        obstruction} \item{stratum}{stratum identifier}
  \item{total}{number of animals observed in each independently
        sighted group [same as GroupSize]} 
  \item{subunit}{a numeric vector, Plot ID} 
}
}
\source{
Patti Happe (Patti_Happe@nps.gov)
}
\description{
Mountain Goat Survey Data from Olympic National park collected in 2004
}
\examples{

data(gdat)

}
\references{
Jenkins, K. J., Happe, P.J., Beirne, K.F, Hoffman, R.A.,
Griffin, P.C., Baccus, W. T., and J. Fieberg.  In press.  Recent population
trends in mountain goats in the Olympic mountains.  Northwest Science.
}
\keyword{datasets}
