% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs.m}
\alias{obs.m}
\title{MN moose survey data}
\format{
A data frame with 805 observations on the following 11 variables.
\describe{ 
  \item{year}{year of survey}
  \item{stratum}{stratum identifier} \item{subunit}{sample
       plot ID} 
  \item{total}{number of moose observed}
  \item{cows}{number of cows observed} \item{calves}{number of
       calves observed} 
  \item{bulls}{number of bulls observed}
  \item{unclass}{number of unclassified animals observed (could not
        identify sex/age class)} 
  \item{voc}{ measurement of visual
       obstruction} 
  \item{grpsize}{group size (cluster size)} 
}
}
\description{
Operational survey data for moose in MN (during years 2004-2007).  Each
record corresponds to an independently sighted group of moose, with variables
that capture individual covariates (used in the detection model) as well as
plot-level information (stratum identifier, sampling probability, etc).
}
\examples{

data(obs.m) 
obs.m[1:5, ]

}
\references{
Giudice, J H. and Fieberg, J. and Lenarz, M. S.  2012. Spending
Degrees of Freedom in a Poor Economy: A Case Study of Building a
Sightability Model for Moose in Northeastern Minnesota.  Journal of Wildlife
Management 76(1):75-87.
}
\keyword{datasets}
