\name{print.sightest}
\alias{print.sightest}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Print method for sightability estimator
}
\description{
Prints fitted sightability model, sampling information, and sightability estimate (with confidence interval)

}
\usage{
  \method{print}{sightest}(x,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Sightability object, output from call to Sight.Est function. 
}
  \item{...}{arguments to be passed to or from other methods}
}
\author{
John Fieberg
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
 \code{\link{Sight.Est}}, \code{\link{summary.sightest}} 
}
 
