\name{print.sightest}
\alias{print.sightest}
\alias{print.sightest_ratio}
\title{Print method for sightability estimators
}
\description{
Prints fitted sightability model, sampling information, and sightability estimate (with confidence interval)
}
\usage{
  \method{print}{sightest}(x,...)
}
\arguments{
  \item{x}{
   Sightability object, output from call to Sight.Est() or Sight.Est.Ratio() functions. 
}
  \item{...}{arguments to be passed to or from other methods}
}
\author{
John Fieberg and Carl James schwarz
}

\seealso{
 \code{\link{Sight.Est}}, \code{\link{Sight.Est.Ratio}}, \code{\link{summary.sightest}}, \code{\link{summary.sightest_ratio}} 
}
\keyword{print}
 
