% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sightest.R
\name{print.sightest}
\alias{print.sightest}
\alias{print.sightest_ratio}
\title{Print method for sightability estimators}
\usage{
\method{print}{sightest}(x, ...)
}
\arguments{
\item{x}{Sightability object, output from call to Sight.Est() or
Sight.Est.Ratio() functions.}

\item{...}{arguments to be passed to or from other methods}
}
\description{
Prints fitted sightability model, sampling information, and sightability
estimate (with confidence interval)
}
\seealso{
\code{\link{Sight.Est}}, \code{\link{Sight.Est.Ratio}},
\code{\link{summary.sightest}}, \code{\link{summary.sightest_ratio}}
}
\author{
John Fieberg and Carl James Schwarz
}
\keyword{print}
