% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampinfo.m}
\alias{sampinfo.m}
\title{Data set containing sampling information for observation survey of moose in
MN}
\format{
A data frame with 12 observations on the following 5 variables.
\describe{ 
  \item{year}{year of survey}
  \item{stratum}{stratum identifier} 
  \item{Nh}{number of
       population units in stratum h} 
  \item{nh}{number of sample units in
       stratum h} 
}
}
\description{
Data set containing sampling information from a survey of moose in MN
(during years 2004-2007)
}
\examples{

data(sampinfo.m)
sampinfo.m

}
\references{
Giudice, J H. and Fieberg, J. and Lenarz, M. S.  2012. Spending
Degrees of Freedom in a Poor Economy: A Case Study of Building a
Sightability Model for Moose in Northeastern Minnesota.  Journal of Wildlife
Management 76(1):75-87.
}
\keyword{datasets}
