% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sightest.R
\name{summary.sightest}
\alias{summary.sightest}
\alias{summary.sightest_ratio}
\title{Summarize sightability estimator}
\usage{
\method{summary}{sightest}(object, ...)
}
\arguments{
\item{object}{Sightability object, output from call to Sight.Est function.}

\item{...}{arguments to be passed to or from other methods}
}
\value{
\item{Nhat or Ratiohat}{Sightability population estimate}
\item{lcl}{Lower confidence limit} \item{ucl}{Upper confidence limit}
}
\description{
Calculates confidence interval (based on asymptotic [normal or log-normal
assumption])
}
\seealso{
\code{\link{Sight.Est}}, \code{\link{Sight.Est.Ratio}}
}
\author{
John Fieberg and Carl James Schwarz
}
\keyword{summary}
